// frontend/src/api/generated/knowledge-client/api.ts
export const JobJobStateEnum = {
  StageUnspecified: "STAGE_UNSPECIFIED",
  Creating: "CREATING",
  Parsing: "PARSING",
  Succeeded: "SUCCEEDED",
  Failed: "FAILED",
  Cancelled: "CANCELLED",
  PartialSuccess: "PARTIAL_SUCCESS",
  Suspended: "SUSPENDED",
} as const;

export type JobJobStateEnum =
  (typeof JobJobStateEnum)[keyof typeof JobJobStateEnum];